Fix some path issues with reorganization in tests
